% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpareto.R
\name{GPareto}
\alias{GPareto}
\alias{dgpareto}
\alias{pgpareto}
\alias{rgpareto}
\alias{qgpareto}
\title{Generalized Pareto distribution}
\usage{
dgpareto(x, mu, gamma, sigma, log = FALSE)

pgpareto(q, mu, gamma, sigma)

rgpareto(n, mu, gamma, sigma)

qgpareto(p, mu, gamma, sigma)
}
\arguments{
\item{x}{numeric vector}

\item{mu}{location parameter}

\item{gamma}{shape parameter}

\item{sigma}{scale parameter, strictly positive}

\item{log}{logical, whether to return the log-density}

\item{q}{numeric vector of quantiles}

\item{n}{positive integer, the desired number of simulations}

\item{p}{numeric vector of probabilities}
}
\description{
Density, distribution function, quantile function, and random 
  generation for the generalized Pareto distribution.
}
