% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{joinMCMCchains}
\alias{joinMCMCchains}
\title{Join MCMC chains}
\usage{
joinMCMCchains(gfi)
}
\arguments{
\item{gfi}{an output of \code{\link{gfigpd1}} or \code{\link{gfigpd2}} 
containing more than one chain}
}
\value{
A \code{mcmc} object.
}
\description{
Joins multiple MCMC chains into a single chain.
}
