% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thresholdEstimate}
\alias{thresholdEstimate}
\title{Threshold estimate}
\usage{
thresholdEstimate(gfi)
}
\arguments{
\item{gfi}{an output of \code{\link{gfigpd2}}}
}
\value{
The estimated threshold.
}
\description{
Returns the estimate of the threshold.
}
